+++ src/client/game/world/World.ts
@@ -1,0 +1,12 @@
+/**
+ * Handler for all in-game logic.
+ */
+export class World {
+
+  /**
+   * Asynchronously load a world with Babylon.
+   */
+  async loadWorld() {
+
+  }
+}